chore: reorganize and rename components for better structure
