+++ src/components/Games/story-game/completion-screen.jsx
