+++ src/components/Games/story-game/draggable-word.jsx
