+++ src/components/Games/story-game/drop-zone.jsx
