+++ src/components/layout/ui/connect-wallet-button.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Wallet } from 'lucide-react';
-import { useWallet } from '../../context/WalletContext'; 
-import { truncateAddress } from '../../utils/helpers';
+import { useWallet } from '../../../context/WalletContext'; 
+import { truncateAddress } from '../../../utils/helpers';
 
 const ConnectWalletButton = () => {
   const { address, connectWallet, disconnectWallet } = useWallet();
