+++ src/components/learning-path/category-card.jsx
