+++ src/components/learning-path/learning-module-details.jsx
