+++ src/components/practices/exercises/questions-section.jsx
