+++ src/components/practices/mock-data-practices/qestion.js
