+++ src/pages/Certification.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import CertificationBanner from './../components/certification/CertificationBanner'; 
-import LearningModule from './../components/certification/LearningModule'; 
+import CertificationBanner from '../components/english-level/english-banner'; 
+import LearningModule from '../components/learning-path/learning-module'; 
 import learningModuleImg from "../assets/learning_module.png"; 
 import certificationBanner from "../assets/certification_banner.png";
