+++ src/pages/FillInTheBlanksQuizPage.jsx
@@ -1,4 +1,4 @@
-import FillInTheBlanksQuiz from '../components/quiz/FillInTheBlanksQuiz';
+import FillInTheBlanksQuiz from '../components/practices/quiz/fill-in-the-blanks-quiz';
 
 export default function FillInTheBlanksQuizPage() {
   return (
