+++ src/pages/ModuleDetails.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useLocation } from "react-router-dom";
 import { BookOpen, FilePlus } from "lucide-react";
-import Sidebar from "../components/layout/Sidebar";
+import Sidebar from "../components/layout/sidebar";
 import certificationBanner from "../assets/certification_banner.png"; // Imagen predeterminada
 
 const ModuleDetails = () => {
